7176907: additional warnings cleanup in java.util, java.util.regexp, java.util.zip

+++ jdk/src/share/classes/java/util/zip/GZIPInputStream.java
@@ -252,7 +252,7 @@ class GZIPInputStream extends InflaterInputStream {
      */
     private int readUShort(InputStream in) throws IOException {
         int b = readUByte(in);
-        return ((int)readUByte(in) << 8) | b;
+        return (readUByte(in) << 8) | b;
     }
 
     /*
